Fix paramsBoostedFields and paramsFilterFields
According to documentation: http://actionml.com/docs/ur_config
fields: optional, default = none. An array of ... . Positive biases are boosts any negative number will filter out any results that do not contain the values in the field name.

When adding fields to algorithm parameters with a negative bias (-1) they will not be filtered, but instead added as boosted fields with a negative value. Providing a positive number will cause this to be a filter.

If you look at the annotations, someone has tried to fix this but instead of swapping the ">" and "<", he moved "=" instead.
I have corrected this issue